+++ CHANGELOG.md
@@ -13,6 +13,8 @@
 * add chromecast as NUHF SDK
 * add ActiveJammer shadow method
 * cleaning up Active jammer code mess
+* removed click artifacts from NUHF tone active jamming
+* add sine/sqr/saw waveforms
 
 ### 4.4.2 (2020-8-26)
 * upgrade build target to API 29, Google Play comply